Use of print function arguments on Python 2.7 requires future
Previously on af1d38b9f, the future import was removed, but that was a
mistake given that the use of the print function here includes
arguments such as "end".

While at it, let's use positional arguments to print(), which is
a good practice, instead of formatting the string and passing to it.

Signed-off-by: Cleber Rosa <crosa@redhat.com>